优化 Apple ID 注册作业和 Apple 服务
- 更新 Apple ID 注册作业，增加对邮箱状态的检查，确保邮箱可用且未失败才进行注册
- 重构 Apple 服务中的连接器配置逻辑，简化代码并提高可读性
- 移除不必要的属性和方法，优化代码结构
- 更新 AppleBuilder，移除不再使用的 basePath 设置

+++ app/Jobs/RegisterAppleIdJob.php
@@ -81,7 +81,7 @@ class RegisterAppleIdJob implements ShouldQueue, ShouldBeUnique
     {
         try {
 
-            if ($this->email->status->value !== EmailStatus::AVAILABLE->value) {
+            if ($this->email->status->value !== EmailStatus::AVAILABLE->value && $this->email->status->value !== EmailStatus::FAILED->value) {
                 Log::warning("Job for {$this->email->email} skipped: Not available.");
                 $this->delete(); // 删除重复的 Job
                 return;
